updated item categories

+++ app/Http/Controllers/ItemCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\AssetCategory;
 use App\Models\ItemCategory;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
 use League\Fractal\Resource\Item;
 use Yajra\DataTables\Facades\DataTables;
 
@@ -41,14 +42,21 @@ class ItemCategoryController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'asset_category_id' => 'required',
-            'item_category_name' => 'required',
-        ]);
-        ItemCategory::create($request->all());
-        return response()->json([
-            'status' => true,
-        ], 200);
+        try {
+            $this->validate($request, [
+                'asset_category_id' => 'required',
+                'item_category_name' => 'required',
+            ]);
+            ItemCategory::create($request->all());
+            return response()->json([
+                'status' => true,
+            ], 200);
+        } catch (\Throwable $th) {
+            Log::error($th->getMessage());
+            return response()->json([
+                'status' => false,
+            ], 500);
+        }
     }
 
     /**
@@ -71,14 +79,21 @@ class ItemCategoryController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $this->validate($request, [
-            'asset_category_id' => 'required',
-            'item_category_name' => 'required',
-        ]);
-        ItemCategory::where('item_category_id', $id)->update($request->all());
-        return response()->json([
-            'status' => true,
-        ], 200);
+        try {
+            $this->validate($request, [
+                'asset_category_id' => 'required',
+                'item_category_name' => 'required',
+            ]);
+            ItemCategory::where('item_category_id', $id)->update($request->all());
+            return response()->json([
+                'status' => true,
+            ], 200);
+        } catch (\Throwable $th) {
+            Log::error($th->getMessage());
+            return response()->json([
+                'status' => false,
+            ], 500);
+        }
     }
 
     /**
@@ -86,9 +101,21 @@ class ItemCategoryController extends Controller
      */
     public function destroy(string $id)
     {
-        ItemCategory::where('item_category_id', $id)->delete();
-        return response()->json([
-            'status' => true,
-        ], 200);
+        try {
+            if ($id == null) {
+                return response()->json([
+                    'status' => false,
+                ], 500);
+            }
+            ItemCategory::where('item_category_id', $id)->delete();
+            return response()->json([
+                'status' => true,
+            ], 200);
+        } catch (\Throwable $th) {
+            Log::error($th->getMessage());
+            return response()->json([
+                'status' => false,
+            ], 500);
+        }
     }
 }
